+++ flatlaf-demo/src/main/java/com/formdev/flatlaf/demo/intellijthemes/IJThemesManager.java
@@ -56,13 +56,16 @@ class IJThemesManager
 			String resourceName = e.getKey();
 			Map<String, String> value = (Map<String, String>) e.getValue();
 			String name = value.get( "name" );
+			boolean discontinued = Boolean.parseBoolean( value.get( "discontinued" ) );
 			boolean dark = Boolean.parseBoolean( value.get( "dark" ) );
 			String license = value.get( "license" );
 			String licenseFile = value.get( "licenseFile" );
+			String pluginUrl = value.get( "pluginUrl" );
 			String sourceCodeUrl = value.get( "sourceCodeUrl" );
 			String sourceCodePath = value.get( "sourceCodePath" );
 
-			bundledThemes.add( new IJThemeInfo( name, resourceName, dark, license, licenseFile, sourceCodeUrl, sourceCodePath, null, null ) );
+			bundledThemes.add( new IJThemeInfo( name, resourceName, discontinued, dark,
+				license, licenseFile, pluginUrl, sourceCodeUrl, sourceCodePath, null, null ) );
 		}
 	}
 
@@ -85,7 +88,7 @@ class IJThemesManager
 			String name = fname.endsWith( ".properties" )
 				? StringUtils.removeTrailing( fname, ".properties" )
 				: StringUtils.removeTrailing( fname, ".theme.json" );
-			moreThemes.add( new IJThemeInfo( name, null, false, null, null, null, null, f, null ) );
+			moreThemes.add( new IJThemeInfo( name, null, false, false, null, null, null, null, null, f, null ) );
 			lastModifiedMap.put( f, f.lastModified() );
 		}
 	}